+++ project/src/components/nav-user/nav-user.tsx
@@ -31,6 +31,7 @@ function NavUser(): JSX.Element {
         </ul>
       );
     }
+
     return (
       <ul className="header__nav-list">
         <li className="header__nav-item user">
