+++ project/src/index.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
-import {ToastContainer} from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 import App from './components/app/app';
 import {store} from './store';
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <ToastContainer />
       <App />
     </Provider>
   </React.StrictMode>,
