+++ project/src/store/action.ts
@@ -5,3 +5,7 @@ import OfferType from '../types/offer';
 export const addOffers = createAction<{offers: OfferType[]}>('data/loadOffers');
 
 export const requireAuthorization = createAction<AuthorizationStatus>('user/requireAuthorization');
+
+export const setOffer = createAction<{activeOffer: Record<string, unknown>}>('data/setOffer');
+
+export const setLoading = createAction<{isLoading: true | false}>('data/setLoading');
