+++ project/src/types/data.ts
@@ -2,3 +2,7 @@ export type AuthData = {
   login: string;
   password: string;
 };
+
+export type RoomData = {
+  id: string | undefined;
+};
